Generate preseed content for certificates
Extend plugin manager to get preseed content
from plugins.
Add preseed content for CA plugin.
Modify generate_deployment_preseed to add
preseed content from plugins.
